+++ src/Emulator/Main/Peripherals/IHasDivisibleFrequency.cs
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2010-2023 Antmicro
+// Copyright (c) 2010-2026 Antmicro
 //
 // This file is licensed under the MIT License.
 // Full license text is available in 'licenses/MIT.txt'.
@@ -9,6 +9,6 @@ namespace Antmicro.Renode.Peripherals
 {
     public interface IHasDivisibleFrequency : IHasFrequency
     {
-        int Divider { get; set; }
+        ulong Divider { get; set; }
     }
 }
