+++ src/Emulator/Peripherals/Peripherals/Timers/EFR32xG2_RTCC.cs
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2010-2023 Antmicro
+// Copyright (c) 2010-2026 Antmicro
 //
 // This file is licensed under the MIT License.
 // Full license text is available in 'licenses/MIT.txt'.
@@ -16,7 +16,7 @@ namespace Antmicro.Renode.Peripherals.Timers
 {
     public class EFR32xG2_RTCC : BasicDoubleWordPeripheral, IKnownSize
     {
-        public EFR32xG2_RTCC(IMachine machine, long frequency) : base(machine)
+        public EFR32xG2_RTCC(IMachine machine, ulong frequency) : base(machine)
         {
             IRQ = new GPIO();
             interruptManager = new InterruptManager<Interrupt>(this);
@@ -112,7 +112,7 @@ namespace Antmicro.Renode.Peripherals.Timers
                 .WithTaggedFlag("CNTCCV1TOP", 2)
                 .WithTaggedFlag("CNTTICK", 3)
                 .WithValueField(4, 4,
-                    writeCallback: (_, value) => innerTimer.Prescaler = (int)Math.Pow(2, value),
+                    writeCallback: (_, value) => innerTimer.Prescaler = (ulong)Math.Pow(2, value),
                     valueProviderCallback: _ => (uint)Math.Log(innerTimer.Prescaler, 2),
                     name: "CNTPRESC")
                 .WithReservedBits(8, 24);
